Adapt Prohibited to the new Menu class
Added Parcelable implement for Menu and Meal class so that they can be transmitted with an intent. Also tried passing a menu to Prohibited.java through a Menu object instead of a group of "C-style" variables.
Be noted that Allowed.java hasn't been adapted to the new way yet, which will cause a crash.  Do not use it.