Replace Parcelable with Serializable
Serializable interface has a better adaptability for SharedPreferences than Parcelable, and is thus more promising for building a database with SharedPreferences in future developments. But we still need to figure out how AllowedList Activity is going to pass data back to the previous Acticity before sending them to the school server.